requirements.txt
streamlit
openai
langchain
python-dotenv
pandas
tiktoken
PyPDF2
chromadb